Refactoring to prepare for alternative task_graph snapshot samplers and a cleaner task graph API.
