+++ agents_playground/ui/main_frame.py
@@ -48,7 +48,7 @@ class LandscapeEditorModes(IntEnum):
 class MainFrame(wx.Frame):
     @log_call
     def __init__(
-        self, sim_path: str | None = None, capture_task_graph_snapshot: bool = False
+        self, sim_path: str | None = None
     ) -> None:
         """
         Create a new Simulation Frame.
@@ -57,7 +57,6 @@ class MainFrame(wx.Frame):
           - sim_path: The path to a simulation to run.
         """
         super().__init__(None, title="The Agent's Playground")
-        self._capture_task_graph_snapshot = capture_task_graph_snapshot
         self._build_ui()
         self._active_simulation: MaybeMutator[SimulationLike] = NothingMutator()
         self._active_landscape: MaybeMutator[WebGPULandscapeEditor] = NothingMutator()
@@ -294,8 +293,7 @@ class MainFrame(wx.Frame):
         # )
         return TaskDrivenSimulation(
             canvas=self.canvas,
-            scene_file=scene_file,
-            capture_task_graph_snapshot=self._capture_task_graph_snapshot,
+            scene_file=scene_file
         )
 
     def _launch_landscape_editor(
